GETP-168 feat: 피플 조회 기능의 피플 좋아요 수 조회 구현 (#111)
* GETP-168 refactor: 좋아요 / 좋아요 취소 서비스를 추상화

* GETP-168 remove: 현재 의미 없는 응용 레이어 테스트 삭제

* GETP-168 rename: 피플 좋아요 애그리거트를 like 패키지로 이동

* GETP-168 feat: 피플 조회 기능의 피플 좋아요 수 조회 구현

* GETP-168 refactor: 프로젝트 좋아요 / 좋아요 취소 서비스를 추상화에 대한 구현체로 변경

+++ src/main/java/es/princip/getp/domain/people/query/dto/people/CardPeopleResponse.java
@@ -8,8 +8,8 @@ public record CardPeopleResponse(
     String nickname,
     String profileImageUri,
     PeopleType peopleType,
-    Integer completedProjectsCount,
-    Integer likesCount,
+    long completedProjectsCount,
+    long likesCount,
     CardPeopleProfileResponse profile
 ) {
 }
